backend/requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
dnspython==1.16.0
fastapi==0.68.0
fastapi-utils==0.2.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
motor==2.4.0
multidict==5.1.0
nest-asyncio==1.5.1
pydantic==1.8.2
pymongo==3.12.0
PyYAML==5.4.1
requests==2.26.0
SQLAlchemy==1.4.23
starlette==0.14.2
tabulate==0.8.9
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
yarl==1.6.3
